api/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.2
fastapi-cli==0.0.5
GeoAlchemy2==0.15.2
geopandas==1.0.1
geopandas-postgis==0.2.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
numpy==2.1.2
packaging==24.1
pandas==2.2.3
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.3
psycopg2==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyogrio==0.10.0
pyproj==3.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
rich==13.9.2
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.40.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
supabase
passlib[bcrypt]
python-jose